feat: Add endpoint for reading RFID
This commit adds a new endpoint `/rfid/read` to the API that allows users to read RFID data from keyboards. It uses the `read_rfid_from_keyboards` function with a timeout of 30 seconds. If no RFID data is found within the timeout period, a 404 error is returned. This feature enhances the RFID functionality and improves usability.